refresh token rotation and redis storage for token ids
refresh token rotation and automatic reuse detection.

when a new access token is request via the usage of a refresh token, the
refresh token mechanism is started. verify the refresh token, extract the token
id from the payload, validate the token id from
`RefreshTokenIdsStorage`, it checks if the token id exists in a Redis
store, if it is valid, then invalidate it and generate a new token
pairs.

in the `generateToken`, it first generate a random UUID as an identifier
for a refresh token, then sign a tokens, one for an access token
containing a payload of a user id and an email, and a one for a refresh
token with only the UUID inside its payloads. Once signing the tokens,
leverage Redis to store the refresh token id in the form of
`user-{userId} ---> {REFRESH_TOKEN_ID}`.